requirements.txt
fitbit
pandas
python-dotenv
requests
